+++ R/rix.R
@@ -117,18 +117,15 @@
 #'   the Nix revision closest to that date, by setting `r_ver = "3.1.0"`, which
 #'   was the version of R current at the time. This ensures that Nix builds a
 #'   completely coherent environment. For security purposes, users that wish to
-#'   install packages from GitHub/Gitlab or from the CRAN archives must provide
+#'   install packages from GitHub/GitLab or from the CRAN archives must provide
 #'   a security hash for each package. `{rix}` automatically precomputes this
 #'   hash for the source directory of R packages from GitHub/Gitlab or from the
 #'   CRAN archives, to make sure the expected trusted sources that match the
-#'   precomputed hashes in the `default.nix` are downloaded. If Nix is
-#'   available, then the hash will be computed on the user's machine, however,
-#'   if Nix is not available, then the hash gets computed on a server that we
-#'   set up for this purposes. This server then returns the security hash as
-#'   well as the dependencies of the packages. It is possible to control this
-#'   behaviour using `options(rix.sri_hash=x)`, where `x` is one of "check_nix"
-#'   (the default), "locally" (use the local Nix installation) or "api_server"
-#'   (use the remote server to compute and return the hash).
+#'   precomputed hashes in the `default.nix` are downloaded, but only if Nix
+#'   is installed. If you need to generate an expression with such packages,
+#'   but are working on a system where you can't install Nix, consider generating
+#'   the expression using a continuous integration service, such as GitHub
+#'   Actions.
 #'
 #'   Note that installing packages from Git or old versions using the `"@"`
 #'   notation or local packages, does not leverage Nix's capabilities for
@@ -190,38 +187,57 @@
 #'   skip_post_processing = FALSE
 #' )
 #' }
-rix <- function(r_ver = NULL,
-                date = NULL,
-                r_pkgs = NULL,
-                system_pkgs = NULL,
-                git_pkgs = NULL,
-                local_r_pkgs = NULL,
-                tex_pkgs = NULL,
-                ide = "none",
-                project_path,
-                overwrite = FALSE,
-                print = FALSE,
-                message_type = "simple",
-                shell_hook = NULL,
-                skip_post_processing = FALSE) {
-  message_type <- match.arg(message_type,
+rix <- function(
+  r_ver = NULL,
+  date = NULL,
+  r_pkgs = NULL,
+  system_pkgs = NULL,
+  git_pkgs = NULL,
+  local_r_pkgs = NULL,
+  tex_pkgs = NULL,
+  ide = "none",
+  project_path,
+  overwrite = FALSE,
+  print = FALSE,
+  message_type = "simple",
+  shell_hook = NULL,
+  skip_post_processing = FALSE
+) {
+  message_type <- match.arg(
+    message_type,
     choices = c("quiet", "simple", "verbose")
   )
 
   if (ide == "other") {
-    stop("ide = 'other' has been deprecated in favour of ide = 'none' as of version 0.15.0.")
+    stop(
+      "ide = 'other' has been deprecated in favour of ide = 'none' as of version 0.15.0."
+    )
   } else if (ide == "code") {
-    warning("The behaviour of the 'ide' argument changed since version 0.15.0; we highly recommend reading this vignette: https://docs.ropensci.org/rix/articles/e-configuring-ide.html if you want to use VS Code.")
-  } else if (!(ide %in% c(
-    "none", "code", "codium", "positron",
-    "radian", "rstudio", "rserver"
-  ))) {
-    stop("'ide' must be one of 'none', 'code', 'codium', 'positron', 'radian', 'rstudio', 'rserver'")
+    warning(
+      "The behaviour of the 'ide' argument changed since version 0.15.0; we highly recommend reading this vignette: https://docs.ropensci.org/rix/articles/e-configuring-ide.html if you want to use VS Code."
+    )
+  } else if (
+    !(ide %in%
+      c(
+        "none",
+        "code",
+        "codium",
+        "positron",
+        "radian",
+        "rstudio",
+        "rserver"
+      ))
+  ) {
+    stop(
+      "'ide' must be one of 'none', 'code', 'codium', 'positron', 'radian', 'rstudio', 'rserver'"
+    )
   }
 
   if (!is.null(date) && !(date %in% available_dates())) {
     # nolint start: line_length_linter
-    stop("The provided date is not available.\nRun available_dates() to see which dates are available.")
+    stop(
+      "The provided date is not available.\nRun available_dates() to see which dates are available."
+    )
     # nolint end
   }
 
@@ -230,12 +246,21 @@ rix <- function(r_ver = NULL,
   }
 
   if (is.null(date) && r_ver == "latest") {
-    stop("'latest' was deprecated in favour of 'latest-upstream' as of version 0.14.0.")
+    stop(
+      "'latest' was deprecated in favour of 'latest-upstream' as of version 0.14.0."
+    )
   }
 
   if (
     !(message_type %in% c("simple", "quiet")) &&
-      r_ver %in% c("bleeding-edge", "frozen-edge", "r-devel", "bioc-devel", "r-devel-bioc-devel")
+      r_ver %in%
+        c(
+          "bleeding-edge",
+          "frozen-edge",
+          "r-devel",
+          "bioc-devel",
+          "r-devel-bioc-devel"
+        )
   ) {
     warning(
       "You chose 'bleeding-edge', 'frozen-edge', 'r-devel', 'bioc-devel' or 'r-devel-bioc-devel'
@@ -246,7 +271,9 @@ before continuing."
   }
 
   if (
-    identical(ide, "rstudio") && is.null(r_pkgs) && is.null(git_pkgs) &&
+    identical(ide, "rstudio") &&
+      is.null(r_pkgs) &&
+      is.null(git_pkgs) &&
       is.null(local_r_pkgs)
   ) {
     stop(
@@ -266,7 +293,8 @@ before continuing."
   )
 
   if (
-    message_type != "quiet" && Sys.info()["sysname"] == "Darwin" &&
+    message_type != "quiet" &&
+      Sys.info()["sysname"] == "Darwin" &&
       ide == "rstudio"
   ) {
     warning(
@@ -303,9 +331,7 @@ for more details."
   cran_pkgs <- get_rpkgs(r_pkgs, ide)
 
   # If there are R packages, passes the string "rpkgs" to buildInputs
-  flag_rpkgs <- if (
-    is.null(cran_pkgs$rPackages) || cran_pkgs$rPackages == ""
-  ) {
+  flag_rpkgs <- if (is.null(cran_pkgs$rPackages) || cran_pkgs$rPackages == "") {
     ""
   } else {
     "rpkgs"
@@ -353,7 +379,8 @@ for more details."
 
   # If there are wrapped packages (for example for RStudio), passes the "wrapped_pkgs"
   # to buildInputs
-  flag_wrapper <- if (ide %in% names(attrib) && flag_rpkgs != "") "wrapped_pkgs" else ""
+  flag_wrapper <- if (ide %in% names(attrib) && flag_rpkgs != "")
+    "wrapped_pkgs" else ""
 
   # Correctly formats shellHook for Nix's mkShell
   shell_hook <- if (!is.null(shell_hook) && nzchar(shell_hook)) {
@@ -370,14 +397,28 @@ for more details."
       ide
     ),
     generate_rpkgs(cran_pkgs$rPackages, flag_rpkgs),
-    generate_git_archived_pkgs(git_pkgs, cran_pkgs$archive_pkgs, flag_git_archive),
+    generate_git_archived_pkgs(
+      git_pkgs,
+      cran_pkgs$archive_pkgs,
+      flag_git_archive
+    ),
     generate_tex_pkgs(tex_pkgs),
     generate_local_r_pkgs(local_r_pkgs, flag_local_r_pkgs),
     generate_system_pkgs(system_pkgs, r_pkgs, ide),
-    generate_wrapped_pkgs(ide, attrib, flag_git_archive, flag_rpkgs, flag_local_r_pkgs),
+    generate_wrapped_pkgs(
+      ide,
+      attrib,
+      flag_git_archive,
+      flag_rpkgs,
+      flag_local_r_pkgs
+    ),
     generate_shell(
-      flag_git_archive, flag_rpkgs, flag_tex_pkgs,
-      flag_local_r_pkgs, flag_wrapper, shell_hook
+      flag_git_archive,
+      flag_rpkgs,
+      flag_tex_pkgs,
+      flag_local_r_pkgs,
+      flag_wrapper,
+      shell_hook
     ),
     collapse = "\n"
   )
@@ -387,7 +428,11 @@ for more details."
   default.nix <- strsplit(default.nix, split = "\n")[[1]]
 
   # Remove potential duplicates
-  default.nix <- post_processing(default.nix, flag_git_archive, skip_post_processing)
+  default.nix <- post_processing(
+    default.nix,
+    flag_git_archive,
+    skip_post_processing
+  )
 
   if (print) {
     print(default.nix)
@@ -400,15 +445,18 @@ for more details."
     con <- file(default.nix_path, open = "wb", encoding = "native.enc")
     on.exit(close(con))
 
-
     writeLines(enc2utf8(default.nix), con = con, useBytes = TRUE)
 
     if (file.exists(.Rprofile_path)) {
-      if (!any(grepl(
-        "File generated by `rix::rix_init()",
-        readLines(.Rprofile_path)
-      ))) {
-        if (message_type != "quiet" && identical(Sys.getenv("TESTTHAT"), "false")) {
+      if (
+        !any(grepl(
+          "File generated by `rix::rix_init()",
+          readLines(.Rprofile_path)
+        ))
+      ) {
+        if (
+          message_type != "quiet" && identical(Sys.getenv("TESTTHAT"), "false")
+        ) {
           message("\n\n### Successfully generated `default.nix` ###\n\n")
         }
         warning(
@@ -417,7 +465,9 @@ for more details."
           "to ensure correct functioning of your Nix environment. ###\n\n"
         )
       } else {
-        if (message_type != "quiet"  && identical(Sys.getenv("TESTTHAT"), "false")) {
+        if (
+          message_type != "quiet" && identical(Sys.getenv("TESTTHAT"), "false")
+        ) {
           message(
             sprintf(
               "\n\n### Successfully generated `default.nix` in %s. ",
@@ -433,13 +483,15 @@ for more details."
         rprofile_action = "create_missing",
         # 'verbose' is too chatty for rix()
         # hence why it's transformed to "simple"
-        message_type = ifelse(message_type == "verbose",
-          "simple", message_type
-        )
+        message_type = ifelse(message_type == "verbose", "simple", message_type)
       )
 
-      if (message_type != "quiet" && identical(Sys.getenv("TESTTHAT"), "false")) {
-        message("\n\n### Successfully generated `default.nix` and `.Rprofile` ###\n\n")
+      if (
+        message_type != "quiet" && identical(Sys.getenv("TESTTHAT"), "false")
+      ) {
+        message(
+          "\n\n### Successfully generated `default.nix` and `.Rprofile` ###\n\n"
+        )
       }
     }
   } else {
@@ -450,35 +502,37 @@ for more details."
     }
     stop(
       paste0(
-        "`default.nix` exists in ", project_path,
+        "`default.nix` exists in ",
+        project_path,
         ". Set `overwrite == TRUE` to overwrite."
       )
     )
   }
 
   on.exit(close(con))
-
-
 }
 
 
 #' @noRd
-post_processing <- function(default.nix, flag_git_archive, skip_post_processing){
-
+post_processing <- function(
+  default.nix,
+  flag_git_archive,
+  skip_post_processing
+) {
   # Remove potential duplicates
   do_processing <- if (flag_git_archive == "") {
-                     FALSE
-                   } else {
-                     TRUE
-                   }
+    FALSE
+  } else {
+    TRUE
+  }
 
   # only do post processing if there are git packages
   # or if skip_post_processing is TRUE
-  if (all(c(do_processing, !skip_post_processing))){
+  if (all(c(do_processing, !skip_post_processing))) {
     out <- remove_duplicate_entries(default.nix) |>
       remove_empty_lines()
   } else {
-    out <-default.nix
+    out <- default.nix
   }
 
   out